Create Reflection
Throughout this entire project 2 progress, I gained a lot of new knowledge. I learn how to do analysis and coding using R programming. Other than that, it was fun doing projects in a groups, where all of us need to coordinate with each other to create a good result where it satisfied everyone.
Beyond the technical skills, this experience highlighted the importance of teamwork, communication, and collective problem-solving.Overall, this project was more than just an academic task, it was a journey of personal and professional growth. Combining technical learning with teamwork was rewarding. I feel well-prepared for future projects and confident that the skills and insights I've gained will be invaluable in my career.